requirements.txt
alembic==1.11.3
annotated-types==0.5.0
asyncpg==0.28.0
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
